+++ src/app/(canvas)/canvas/workspaces/[workspaceId]/page.tsx
@@ -1,11 +1,18 @@
+import { auth } from "@/lib/auth";
 import { Banner } from "./banner";
 import { ProjectsSection } from "./projects-section";
 import { TemplatesSection } from "./templates-section";
+import { headers } from "next/headers";
+import { redirect } from "next/navigation";
 
-export default function WorkspaceIdPage() {
+export default async function WorkspaceIdPage() {
+  const current = await auth.api.getSession({
+    headers: await headers(),
+  });
+  if (!current) redirect("/sign-in");
   return (
     <div className="max-w-7xl mx-auto space-y-10">
-      <Banner/>
+      <Banner />
       <TemplatesSection />
       <ProjectsSection />
     </div>
